Moved models to separate files
