+++ PortfolioApp/Model/UsersModels/AddressModels.swift
@@ -1,0 +1,8 @@
+import Foundation
+
+struct Address: Decodable {
+    let city: String
+    let street: String
+    let suite: String
+    let zipcode: String
+}
