+++ PortfolioApp/Model/UsersModels/CompanyModels.swift
@@ -1,0 +1,5 @@
+import Foundation
+
+struct Company: Decodable {
+    let name: String
+}
